Refactor to Fluent 2
