+++ lib/presentation/game/thumb.dart
